Make design more flexible
Since workload interpretation is a function of the role itself, there needs to be a way to define
survey questions based on a given job role. Templates classes wouldn't
work because the number of data members themselves are
non-deterministic. Since the application depends on user entries, all of
the user-submitted data are text strings. I use vectors of strings to
store the questions and the responses, respectively. This enables the use
of one data structure with a consistent interface to handle an
unspecified number of questions/metrics. The values are later converted
to the requisite types as needed (metric questions are assumed to be
doubles, for instance).

Removes unnecessary classes.